refactor: share dashboard client api
